requirements.txt
streamlit>=1.30.0
torch>=2.0.1
torchvision>=0.15.2
opencv-python-headless>=4.8.0.76
Pillow>=10.0.0
numpy>=1.24.3